requirements.txt
Flask~=3.0.3
Flask-Cors~=5.0.0
aiogram~=3.12.0
APScheduler~=3.10.4
python-dateutil~=2.9.0.post0